Article refactor - Likes, Views and Articles refactor

+++ api/users/users.js
@@ -1,5 +1,5 @@
 const { issuer, panel } = require('../../config/environment')
-const { s3, bucket } = require('../../config/aws/s3')
+const { s3, bucket, getBucketKeyFromUrl } = require('../../config/aws/s3')
 const multer = require('../../config/serialization/multers3')
 
 const uploadImg = multer.single('profilePhoto')
@@ -778,10 +778,20 @@ module.exports = app => {
           return res.status(stack.code).send(stack)
         }
 
+        if (!req.file) {
+          const stack = {
+            code: 400,
+            name: 'profilePhoto',
+            description: 'É necessário selecionar uma imagem'
+          }
+
+          return res.status(stack.code).send(stack)
+        }
+
         const currentProfilePhoto = user.profilePhoto || null
 
         if (currentProfilePhoto) {
-          const { status, key, error } = getBucketObjKeyFromUrl(currentProfilePhoto)
+          const { status, key, error } = getBucketKeyFromUrl(currentProfilePhoto)
           if (!status) return
 
           s3.deleteObject({ Bucket: bucket, Key: key }, (err, data) => {
@@ -801,29 +811,6 @@ module.exports = app => {
     } catch (error) {
       const stack = userError(error)
       return res.status(stack.code).send(stack)
-    }
-  }
-
-  /**
-   * @function
-   * @description Get S3 bucket object key from their url.
-   * @param {String} url - S3 Bucket Object url
-   *
-   * @returns An object containing `status` of operation: `true` for successful,
-   *  `false` for fail; `key` containing the object key; `error` containing the stack error,
-   *  if happens.
-   */
-  const getBucketObjKeyFromUrl = url => {
-    try {
-      if (url && typeof url === 'string') {
-        const parts = url.split('/')
-
-        return { status: true, key: parts[3], error: null }
-      }
-
-      throw new Error(`typeof url is ${typeof url}, expected string`)
-    } catch (error) {
-      return { status: false, key: null, error }
     }
   }
 
@@ -855,7 +842,7 @@ module.exports = app => {
         }
       }
 
-      const { status, key, error } = getBucketObjKeyFromUrl(find.profilePhoto)
+      const { status, key, error } = getBucketKeyFromUrl(find.profilePhoto)
       if (!status) throw error
 
       s3.deleteObject({ Bucket: bucket, Key: key }, async (err, data) => {
